+++ src/quick/scenegraph/util/qsgopenglatlastexture_p.h
@@ -103,7 +103,7 @@ public:
 
     void invalidate();
 
-    int comparisonKey() const { return m_texture_id; }
+    qint64 comparisonKey() const { return m_texture_id; }
     int textureId() const;
     void bind(QSGTexture::Filtering filtering);
 
@@ -159,7 +159,7 @@ public:
     TextureBase(AtlasBase *atlas, const QRect &textureRect);
     ~TextureBase();
 
-    int comparisonKey() const override;
+    qint64 comparisonKey() const override;
     int textureId() const override { return m_atlas->textureId(); }
     bool isAtlasTexture() const override { return true; }
